requirements.txt
selenium==4.21.0
schedule==1.2.2
python-dotenv==1.0.1
webdriver-manager==4.0.1
